fix style and structure
